# Conflicts:
#	app/src/main/java/com/sofiamarchinskaya/hw1/Constants.kt
#	app/src/main/java/com/sofiamarchinskaya/hw1/models/entity/Note.kt
#	app/src/main/java/com/sofiamarchinskaya/hw1/view/NotesListFragment.kt
#	app/src/main/java/com/sofiamarchinskaya/hw1/viewmodels/NotesListViewModel.kt

+++ app/src/main/java/com/sofiamarchinskaya/hw1/models/ListWithIndex.kt
@@ -2,4 +2,4 @@ package com.sofiamarchinskaya.hw1.models
 
 import com.sofiamarchinskaya.hw1.models.entity.Note
 
-data class ListWithIndex(val list: List<Note>, val index: Long)
+data class ListWithIndex(val list: List<Note>, val index: Int)
