+++ app/src/main/java/com/sofiamarchinskaya/hw1/viewmodels/NotesPagerViewModel.kt
@@ -9,11 +9,11 @@ import com.sofiamarchinskaya.hw1.models.framework.NoteRepository
 class NotesPagerViewModel(private val repository: NoteRepository) : ViewModel() {
     val listWithIndex = MutableLiveData<ListWithIndex>()
 
-    suspend fun init(id: Long?) =
+    suspend fun init(id: Int?) =
         repository.getAll().collect {
-            listWithIndex.value = ListWithIndex(it, getIndex(id, it).toLong())
+            listWithIndex.value = ListWithIndex(it, getIndex(id, it))
         }
 
-    private fun getIndex(id: Long?, list: List<Note>) =
+    private fun getIndex(id: Int?, list: List<Note>) =
         list.indexOfFirst { it.id == id }
 }
